фикс верстки и вынос функции отправки списка в Util
